Add 10 Min Revalidation, Fix logging, switch to revalidateRoute
