+++ next.config.js
@@ -19,7 +19,7 @@ module.exports = withBundleAnalyzer({
   compiler: {
     removeConsole:
       process.env.NODE_ENV === 'production'
-        ? { exclude: ['error', 'warn'] }
+        ? { exclude: ['error', 'warn', 'log'] }
         : false,
   },
 });
